requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.24.3
python-dotenv==1.0.0
google-search-results==2.4.2
groq==0.4.2
openpyxl==3.1.2
xlsxwriter==3.1.2
python-json-logger==2.0.7
typing-extensions==4.7.1